+++ src/views/window.rs
@@ -678,6 +678,14 @@ impl View for Window {
     fn constrain_to_parent_bounds(&mut self) {
         self.constrain_to_limits();
     }
+
+    fn as_any(&self) -> &dyn std::any::Any {
+        self
+    }
+
+    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
+        self
+    }
 }
 
 /// Builder for creating windows with a fluent API.
